Added several class skeletons.
It's a very rough carving of some of the basics. I still need to flesh
out the controller scheme, but we can use mouse and keyboard for now. We
may have to add our own controller class to do things right, but we'll
see.

I still need to set up the battle game structure as well. That is,
collision, object physics, and controlling characters.