Attacks and Collision Boxes are in the game!
An Attack is a collection of collision boxes.
A collision box has A LOT of properties to help give a diverse range of
attack types.
More properties may have to be added to introduce additional physics.
Some input stuff has changed.
Collisions happen between collision boxes and with player outlines.